+++ include/kernel.h
@@ -34,7 +34,6 @@
 
 #include "config.h"
 #include "mt32pi.h"
-#include "zoneallocator.h"
 
 class CKernel : public CStdlibApp
 {
@@ -62,7 +61,6 @@ protected:
 	CGPIOManager m_GPIOManager;
 
 private:
-	CZoneAllocator m_Allocator;
 	CConfig m_Config;
 	CMT32Pi m_MT32Pi;
 };